cleaned up code in event pages and cards
