+++ src/components/cards/EventCardDesktop.js
@@ -9,30 +9,9 @@ import Heading1 from "../text/Heading1";
 import Heading2 from "../text/Heading2";
 import Circle from "react-simple-shapes";
 import EventEmailModal from "./EventEmailModal"
+import {months, formatTime} from "./eventCardsFeaturedAndCards/commonEventsFuncs"
 
 const theme = CustomTheme;
-
-const months = {
-  0: "January",
-  1: "February",
-  2: "March",
-  3: "April",
-  4: "May",
-  5: "June",
-  6: "July",
-  7: "August",
-  8: "September",
-  9: "October",
-  10: "November",
-  11: "December"
-};
-
-const formatTime = function(hours, min) {
-  let h = hours > 12 ? hours - 12 : hours;
-  let m = min < 10 ? "0" + min.toString() : min.toString();
-  let add = hours > 12 ? "PM" : "AM";
-  return h + ":" + m + add;
-};
 
 const useStyles = makeStyles(() => ({
   card: {
