+++ src/components/events/eventPageDesktop.js
@@ -8,11 +8,6 @@ import { EventCardFeatured, EventCard, EventModal, Template, CustomButton, Title
   from "../";
 import firebase from "../../firebase";
 import Fuse from 'fuse.js';
-import {
-  getTimezoneName, convertUTCToLocal, convertDateToUTC,
-  getOffset, getCurrentLocationForTimeZone, dst, convertTimestampToDate
-}
-  from "../all/TimeFunctions"
 import CustomToolbar from "../events/CalendarToolBar"
 import {CircularProgress} from "@material-ui/core";
 import GridItem from "../material-kit-components/Grid/GridItem.js";
@@ -24,7 +19,7 @@ import LinkedInIcon from '@material-ui/icons/LinkedIn';
 import MailOutlineIcon from '@material-ui/icons/MailOutline';
 import ScrollableAnchor from 'react-scrollable-anchor';
 import { configureAnchors } from 'react-scrollable-anchor';
-import {eventPropStylesShared} from "./SharedEvents";
+import {eventPropStylesShared, convertEventsTime, makeDisplayEvents, isEventShowable} from "./SharedEvents";
 configureAnchors({ offset: -100 });
 
 const localizer = momentLocalizer(moment);
@@ -205,34 +200,6 @@ class EventsPageDesktop extends React.Component {
     this.resetFilter = this.resetFilter.bind(this);
   }
 
-  convertEventsTime(event) {
-    const tzString = event.timezone;
-
-    // Remove redudancy (AKA remove the evidence -.0)
-    event.start_date = event.start_date.split("GMT")[0];
-    event.end_date = event.end_date.split("GMT")[0];
-
-    if (event.timezone !== undefined && event.timezone.includes("$")) {
-      // $ splits time and timezone in the event.timezone field in firebase!
-      const tz = tzString.split("$")[0];
-      const daylightSavings = tzString.split("$")[1] === "true" ? true : false;
-      const offset = getOffset(tz, daylightSavings);
-
-      // First convert the event's time to UTC, assuming the event is in EST time (America/New_York)
-      // America/New_York should be changed to the user's time zone who created the event, if they
-      // Choose to use their time zone rather than EST.
-      const UTCStart = convertDateToUTC(convertTimestampToDate(event.start_date), offset);
-      const UTCEnd = convertDateToUTC(convertTimestampToDate(event.end_date), offset);
-
-      // Second, convert those consts above to user's local time
-      event.start_date = convertUTCToLocal(UTCStart);
-      event.end_date = convertUTCToLocal(UTCEnd);
-      // get timezone to display
-      event.timeZoneGMT = getTimezoneName(getCurrentLocationForTimeZone(), dst());
-    }
-    return event;
-  }
-
   async componentDidMount() {
     await this.getEvents();
     let event = this.props.event
@@ -273,24 +240,6 @@ class EventsPageDesktop extends React.Component {
   //     this.setState({ myEventsList: events, displayEvents: this.makeDisplayEvents(events) });
   //   })
   // }
-
-  makeDisplayEvents(events) {
-    let arr = [];
-    for (let i = 0; i < events.length; i += 1) {
-      let ele = events[i];
-
-      ele.title = ele.event === undefined ? ele.title : ele.event
-      ele.event = ele.title
-
-      if (ele.end_date > new Date()) {
-        arr.push(ele);
-      }
-      if (arr.length === 5) {
-        break;
-      }
-    }
-    return arr;
-  }
 
   makeEventsList(events) {
     return events;
@@ -347,7 +296,7 @@ class EventsPageDesktop extends React.Component {
     if(approvedEvents){
       approvedEventsMap = approvedEvents.docs.map(doc => {
 
-        let event = this.convertEventsTime(doc.data())
+        let event = convertEventsTime(doc.data())
         event["id"] = doc.id
         let today = new Date()
         if ((new Date(event.start_date)) < today && (new Date(event.end_date)) > today) {
@@ -382,7 +331,7 @@ class EventsPageDesktop extends React.Component {
       myEventsList: this.makeEventsList(approvedEventsMap), tagList: this.genTagsList(approvedEventsMap),
       organizationList: this.genOrganizationList(approvedEventsMap),
       permEventsList: approvedEventsMap,
-      displayEvents: this.makeDisplayEvents(approvedEventsMap),
+      displayEvents: makeDisplayEvents(approvedEventsMap),
       loadingEvents: false,
       loadingFeaturedEvents: false
     });
@@ -461,40 +410,6 @@ class EventsPageDesktop extends React.Component {
     this.setState({ open: false, count: 0 });
   }
 
-  eventPropStyles(event, start, end, isSelected) {
-    let style = {
-      backgroundColor: "#2984ce"
-    };
-    let nowStyle = {
-      backgroundColor: "#F3FFEE"
-    }
-    let pastStyle = {
-      backgroundColor: "#BDBDBD"
-    }
-    let popularStyle = {
-      backgroundColor: "#F2F9FD"
-    }
-    let recurringStyle = {
-      backgroundColor: "#FDEEE5"
-    }
-
-    if (event.displayNow) {
-      return { style: nowStyle };
-    }
-    else if (event.displayPast) {
-      return { style: pastStyle };
-    }
-    else if (event.displayPopular) {
-      return { style: popularStyle };
-    }
-    else if (event.displayRecurring) {
-      return { style: recurringStyle };
-    }
-    else {
-      return { style: style };
-    }
-  }
-
   EventDisplay = ({ event }) => {
     return (
     <div style={{ height: "1.2em" }}>
@@ -514,79 +429,6 @@ class EventsPageDesktop extends React.Component {
       newList[tag] = "on"
     }
     this.setState({ mainTagsClicked: newList })
-  }
-
-  isEventShowable(ele) {
-    ele.tagsForFilter = this.state.mainTagsClicked
-    // Handle main tags filter
-    let shouldDisplayRecurring = ele["tagsForFilter"].recurring === "on" ? true : false
-    let shouldDisplayPopular = ele["tagsForFilter"].popular === "on" ? true : false
-    let shouldDisplayPast = ele["tagsForFilter"].past === "on" ? true : false
-    let shouldDisplayNow = ele["tagsForFilter"].now === "on" ? true : false
-
-    ele.displayThisCard = true
-    if (!ele.displayPopular && shouldDisplayPopular)
-      ele.displayThisCard = false
-    if (!ele.displayNow && shouldDisplayNow)
-      ele.displayThisCard = false
-    if (!ele.displayRecurring && shouldDisplayRecurring)
-      ele.displayThisCard = false
-    if (!ele.displayRecurring && shouldDisplayRecurring)
-      ele.displayThisCard = false
-    if (!ele.displayPast && shouldDisplayPast)
-      ele.displayThisCard = false
-    if (ele.displayThisCard && ele.displayPast && !shouldDisplayPast)
-      ele.displayThisCard = false
-
-    // Process regular tag filter
-    let filterPass = true
-    if (ele.displayThisCard === true) {
-      Object.keys(this.state.filterTagsClicked).map(x => {
-        let found = false
-        if (this.state.filterTagsClicked[x] !== undefined) {
-          ele.tags.map(y => {
-            if (x.toLowerCase() === y.toLowerCase()) {
-              found = true
-            }
-          })
-
-          if (found === false) {
-            filterPass = false
-          }
-        }
-      })
-    }
-
-    if (filterPass === false) {
-      ele.displayThisCard = false
-    }
-
-    // Handle Organization
-    if (this.state.clubFilter !== "All") {
-      if (this.state.clubFilter !== ele.name)
-        ele.displayThisCard = false
-    }
-
-    // Handle date
-    if (this.state.dateFilter !== "All") {
-      let d = new Date()
-      const daysApart = Math.abs((ele.start_date.getTime() - d.getTime()) / (3600*24*1000))
-      switch(this.state.dateFilter) {
-        case "This Month Only":
-          if (ele.start_date.getMonth() !== d.getMonth() || ele.start_date.getFullYear() !== d.getFullYear())
-            ele.displayThisCard = false
-        case "Within a Week":
-          if (daysApart > 7)
-            ele.displayThisCard = false
-        case "Within a Month":
-          if (daysApart > 30)
-            ele.displayThisCard = false
-        case "Within 3 Months":
-          if (daysApart > 90)
-            ele.displayThisCard = false
-      }
-    }
-    return ele.displayThisCard
   }
 
   updateOrganization(club) {
@@ -621,7 +463,7 @@ class EventsPageDesktop extends React.Component {
         ele.tags = ['none']
       }
 
-      if (this.isEventShowable(ele)) {
+      if (isEventShowable(ele, this.state.mainTagsClicked, this.state.filterTagsClicked, this.state.clubFilter, this.state.dateFilter)) {
         sizeOfList = sizeOfList + 1
         if (ele.displayNameToggleOff)
           ele.name = "Columbia Virtual Campus"
